[Update] Refactored base models names & added Token model
