+++ app/src/main/java/com/example/authio/models/User.java
@@ -1,21 +1,15 @@
-package com.example.authio.api;
+package com.example.authio.models;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
 // a transcribed model for an HTTP response from the Web service
 // contains 'response' (status), 'username', 'description', and 'email' at most in a JSON format
-public class UserModel {
+public class User {
 
     @SerializedName("id")
     @Expose
     private Integer id;
-
-    @SerializedName("response") // indicates that the field should be serialized to JSON
-    @Expose // indicates the field should be exposed for said JSON serialization
-    private String response; // i.e. the field is part of a JSON format and the name inside is the field value (key)
-    // the value of the JSON key (constructor argument) should be specified from the Web service used
-    // and the response it sends from HTTP
 
     @SerializedName("username")
     @Expose
@@ -29,7 +23,14 @@ public class UserModel {
     @Expose
     private String email;
 
-    public UserModel(Integer id, String response, String username, String description, String email) {
+
+    @SerializedName("response") // indicates that the field should be serialized to JSON
+    @Expose // indicates the field should be exposed for said JSON serialization
+    private String response; // i.e. the field is part of a JSON format and the name inside is the field value (key)
+    // the value of the JSON key (constructor argument) should be specified from the Web service used
+    // and the response it sends from HTTP
+
+    public User(Integer id, String response, String username, String description, String email) {
         this.id = id;
         this.response = response;
         this.username = username;
